requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
plotly
streamlit
ipykernel
google-generativeai
supabase